Moving database initialization to main activity
1104
